Switched old implementation for experimental implementation
