+++ Cargo.toml
@@ -18,14 +18,12 @@ default = ["par-iter"]
 par-iter = ["rayon"]
 
 [dependencies]
-slog = { version = "2.4", features = ["nested-values"] }
-slog-stdlog = "3"
-parking_lot = "0.7"
+slog = { version = "2.5", features = ["nested-values"] }
+slog-stdlog = "4"
+parking_lot = "0.9"
 downcast-rs = "1.0"
 itertools = "0.8"
-names = "0.11.0"
-rayon = { version = "1.1", optional = true }
-fnv = "1.0"
+rayon = { version = "1.2", optional = true }
 derivative = "1"
 smallvec = "0.6"
 
@@ -46,8 +44,4 @@ harness = false
 
 [[bench]]
 name = "transform"
-harness = false
-
-[[bench]]
-name = "transform_experimental"
 harness = false
