+++ src/filter.rs
@@ -1,15 +1,15 @@
-use crate::experimental::storage::ArchetypeData;
-use crate::experimental::storage::ArchetypeId;
-use crate::experimental::storage::ChunkId;
-use crate::experimental::storage::Component;
-use crate::experimental::storage::ComponentStorage;
-use crate::experimental::storage::ComponentTypeId;
-use crate::experimental::storage::ComponentTypes;
-use crate::experimental::storage::SliceVecIter;
-use crate::experimental::storage::Storage;
-use crate::experimental::storage::Tag;
-use crate::experimental::storage::TagTypeId;
-use crate::experimental::storage::TagTypes;
+use crate::storage::ArchetypeData;
+use crate::storage::ArchetypeId;
+use crate::storage::ChunkId;
+use crate::storage::Component;
+use crate::storage::ComponentStorage;
+use crate::storage::ComponentTypeId;
+use crate::storage::ComponentTypes;
+use crate::storage::SliceVecIter;
+use crate::storage::Storage;
+use crate::storage::Tag;
+use crate::storage::TagTypeId;
+use crate::storage::TagTypes;
 use std::collections::HashMap;
 use std::iter::Enumerate;
 use std::iter::Repeat;
