+++ tests/world_api.rs
@@ -1,5 +1,4 @@
 use legion::prelude::*;
-use std::sync::Arc;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 struct Pos(f32, f32, f32);
@@ -18,19 +17,19 @@ struct Static;
 
 #[test]
 fn insert() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
     let shared = (1usize, 2f32, 3u16);
     let components = vec![(4f32, 5u64, 6u16), (4f32, 5u64, 6u16)];
-    let entities = world.insert_from(shared.as_tags(), components);
+    let entities = world.insert(shared, components);
 
     assert_eq!(2, entities.len());
 }
 
 #[test]
 fn get_component() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
     let shared = (Static, Model(5));
@@ -40,16 +39,16 @@ fn get_component() {
     ];
 
     let mut entities: Vec<Entity> = Vec::new();
-    for e in world.insert_from(shared.as_tags(), components.clone()) {
+    for e in world.insert(shared, components.clone()) {
         entities.push(*e);
     }
 
     for (i, e) in entities.iter().enumerate() {
-        match world.component(*e) {
+        match world.get_component(*e) {
             Some(x) => assert_eq!(components.get(i).map(|(x, _)| x), Some(&x as &Pos)),
             None => assert_eq!(components.get(i).map(|(x, _)| x), None),
         }
-        match world.component(*e) {
+        match world.get_component(*e) {
             Some(x) => assert_eq!(components.get(i).map(|(_, x)| x), Some(&x as &Rot)),
             None => assert_eq!(components.get(i).map(|(_, x)| x), None),
         }
@@ -58,17 +57,17 @@ fn get_component() {
 
 #[test]
 fn get_component_wrong_type() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
-    let entity = *world.insert_from((), vec![(0f64,)]).get(0).unwrap();
+    let entity = *world.insert((), vec![(0f64,)]).get(0).unwrap();
 
-    assert_eq!(None, world.component::<i32>(entity));
+    assert!(world.get_component::<i32>(entity).is_none());
 }
 
 #[test]
 fn get_shared() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
     let shared = (Static, Model(5));
@@ -78,32 +77,29 @@ fn get_shared() {
     ];
 
     let mut entities: Vec<Entity> = Vec::new();
-    for e in world.insert_from(shared.as_tags(), components.clone()) {
+    for e in world.insert(shared, components.clone()) {
         entities.push(*e);
     }
 
     for e in entities.iter() {
-        assert_eq!(Some(&Static), world.tag(*e));
-        assert_eq!(Some(&Model(5)), world.tag(*e));
+        assert_eq!(Some(&Static), world.get_tag(*e));
+        assert_eq!(Some(&Model(5)), world.get_tag(*e));
     }
 }
 
 #[test]
 fn get_shared_wrong_type() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
-    let entity = *world
-        .insert_from((Static,).as_tags(), vec![(0f64,)])
-        .get(0)
-        .unwrap();
+    let entity = *world.insert((Static,), vec![(0f64,)]).get(0).unwrap();
 
-    assert_eq!(None, world.tag::<Model>(entity));
+    assert!(world.get_tag::<Model>(entity).is_none());
 }
 
 #[test]
 fn delete() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
     let shared = (Static, Model(5));
@@ -113,7 +109,7 @@ fn delete() {
     ];
 
     let mut entities: Vec<Entity> = Vec::new();
-    for e in world.insert_from(shared.as_tags(), components.clone()) {
+    for e in world.insert(shared, components.clone()) {
         entities.push(*e);
     }
 
@@ -129,7 +125,7 @@ fn delete() {
 
 #[test]
 fn delete_last() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
     let shared = (Static, Model(5));
@@ -139,7 +135,7 @@ fn delete_last() {
     ];
 
     let mut entities: Vec<Entity> = Vec::new();
-    for e in world.insert_from(shared.as_tags(), components.clone()) {
+    for e in world.insert(shared, components.clone()) {
         entities.push(*e);
     }
 
@@ -149,11 +145,11 @@ fn delete_last() {
 
     for (i, e) in entities.iter().take(entities.len() - 1).enumerate() {
         assert_eq!(true, world.is_alive(*e));
-        match world.component(*e) {
+        match world.get_component(*e) {
             Some(x) => assert_eq!(components.get(i).map(|(x, _)| x), Some(&x as &Pos)),
             None => assert_eq!(components.get(i).map(|(x, _)| x), None),
         }
-        match world.component(*e) {
+        match world.get_component(*e) {
             Some(x) => assert_eq!(components.get(i).map(|(_, x)| x), Some(&x as &Rot)),
             None => assert_eq!(components.get(i).map(|(_, x)| x), None),
         }
@@ -162,7 +158,7 @@ fn delete_last() {
 
 #[test]
 fn delete_first() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
     let shared = (Static, Model(5));
@@ -172,7 +168,7 @@ fn delete_first() {
     ];
 
     let mut entities: Vec<Entity> = Vec::new();
-    for e in world.insert_from(shared.as_tags(), components.clone()) {
+    for e in world.insert(shared, components.clone()) {
         entities.push(*e);
     }
 
@@ -183,11 +179,11 @@ fn delete_first() {
 
     for (i, e) in entities.iter().skip(1).enumerate() {
         assert_eq!(true, world.is_alive(*e));
-        match world.component(*e) {
+        match world.get_component(*e) {
             Some(x) => assert_eq!(components.get(i + 1).map(|(x, _)| x), Some(&x as &Pos)),
             None => assert_eq!(components.get(i + 1).map(|(x, _)| x), None),
         }
-        match world.component(*e) {
+        match world.get_component(*e) {
             Some(x) => assert_eq!(components.get(i + 1).map(|(_, x)| x), Some(&x as &Rot)),
             None => assert_eq!(components.get(i + 1).map(|(_, x)| x), None),
         }
@@ -196,7 +192,7 @@ fn delete_first() {
 
 #[test]
 fn merge() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world_1 = universe.create_world();
     let mut world_2 = universe.create_world();
 
@@ -207,12 +203,12 @@ fn merge() {
     ];
 
     let mut world_1_entities: Vec<Entity> = Vec::new();
-    for e in world_1.insert_from(shared.as_tags(), components.clone()) {
+    for e in world_1.insert(shared, components.clone()) {
         world_1_entities.push(*e);
     }
 
     let mut world_2_entities: Vec<Entity> = Vec::new();
-    for e in world_2.insert_from(shared.as_tags(), components.clone()) {
+    for e in world_2.insert(shared, components.clone()) {
         world_2_entities.push(*e);
     }
 
@@ -222,24 +218,24 @@ fn merge() {
         assert!(world_1.is_alive(*e));
 
         let (pos, rot) = components.get(i).unwrap();
-        assert_eq!(pos, &world_1.component(*e).unwrap() as &Pos);
-        assert_eq!(rot, &world_1.component(*e).unwrap() as &Rot);
+        assert_eq!(pos, &world_1.get_component(*e).unwrap() as &Pos);
+        assert_eq!(rot, &world_1.get_component(*e).unwrap() as &Rot);
     }
 }
 
 #[test]
 fn mutate_add_component() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
-    let shared = (Static, Model(5)).as_tags();
+    let shared = (Static, Model(5));
     let components = vec![
         (Pos(1., 2., 3.), Rot(0.1, 0.2, 0.3)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
     ];
 
-    let entities = world.insert_from(shared, components).to_vec();
+    let entities = world.insert(shared, components).to_vec();
 
     let mut query_without_scale = <(Read<Pos>, Read<Rot>)>::query();
     let mut query_with_scale = <(Read<Pos>, Read<Rot>, Read<Scale>)>::query();
@@ -247,9 +243,7 @@ fn mutate_add_component() {
     assert_eq!(3, query_without_scale.iter(&world).count());
     assert_eq!(0, query_with_scale.iter(&world).count());
 
-    world.mutate_entity(*entities.get(1).unwrap(), |e| {
-        e.add_component(Scale(0.5, 0.5, 0.5))
-    });
+    world.add_component(*entities.get(1).unwrap(), Scale(0.5, 0.5, 0.5));
 
     assert_eq!(3, query_without_scale.iter(&world).count());
     assert_eq!(1, query_with_scale.iter(&world).count());
@@ -257,17 +251,17 @@ fn mutate_add_component() {
 
 #[test]
 fn mutate_remove_component() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
-    let shared = (Static, Model(5)).as_tags();
+    let shared = (Static, Model(5));
     let components = vec![
         (Pos(1., 2., 3.), Rot(0.1, 0.2, 0.3)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
     ];
 
-    let entities = world.insert_from(shared, components).to_vec();
+    let entities = world.insert(shared, components).to_vec();
 
     let mut query_without_rot = Read::<Pos>::query().filter(!component::<Rot>());
     let mut query_with_rot = <(Read<Pos>, Read<Rot>)>::query();
@@ -275,9 +269,7 @@ fn mutate_remove_component() {
     assert_eq!(0, query_without_rot.iter(&world).count());
     assert_eq!(3, query_with_rot.iter(&world).count());
 
-    world.mutate_entity(*entities.get(1).unwrap(), |e| {
-        e.remove_component::<Rot>();
-    });
+    world.remove_component::<Rot>(*entities.get(1).unwrap());
 
     assert_eq!(1, query_without_rot.iter(&world).count());
     assert_eq!(2, query_with_rot.iter(&world).count());
@@ -285,17 +277,17 @@ fn mutate_remove_component() {
 
 #[test]
 fn mutate_add_tag() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
-    let shared = (Model(5),).as_tags();
+    let shared = (Model(5),);
     let components = vec![
         (Pos(1., 2., 3.), Rot(0.1, 0.2, 0.3)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
     ];
 
-    let entities = world.insert_from(shared, components).to_vec();
+    let entities = world.insert(shared, components).to_vec();
 
     let mut query_without_static = <(Read<Pos>, Read<Rot>)>::query();
     let mut query_with_static = <(Read<Pos>, Read<Rot>, Tagged<Static>)>::query();
@@ -303,9 +295,7 @@ fn mutate_add_tag() {
     assert_eq!(3, query_without_static.iter(&world).count());
     assert_eq!(0, query_with_static.iter(&world).count());
 
-    world.mutate_entity(*entities.get(1).unwrap(), |e| {
-        e.set_tag(Arc::new(Static));
-    });
+    world.add_tag(*entities.get(1).unwrap(), Static);
 
     assert_eq!(3, query_without_static.iter(&world).count());
     assert_eq!(1, query_with_static.iter(&world).count());
@@ -313,17 +303,17 @@ fn mutate_add_tag() {
 
 #[test]
 fn mutate_remove_tag() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
-    let shared = (Model(5), Static).as_tags();
+    let shared = (Model(5), Static);
     let components = vec![
         (Pos(1., 2., 3.), Rot(0.1, 0.2, 0.3)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
     ];
 
-    let entities = world.insert_from(shared, components).to_vec();
+    let entities = world.insert(shared, components).to_vec();
 
     let mut query_without_static = <(Read<Pos>, Read<Rot>)>::query().filter(!tag::<Static>());
     let mut query_with_static = <(Read<Pos>, Read<Rot>, Tagged<Static>)>::query();
@@ -331,9 +321,7 @@ fn mutate_remove_tag() {
     assert_eq!(0, query_without_static.iter(&world).count());
     assert_eq!(3, query_with_static.iter(&world).count());
 
-    world.mutate_entity(*entities.get(1).unwrap(), |e| {
-        e.remove_tag::<Static>();
-    });
+    world.remove_tag::<Static>(*entities.get(1).unwrap());
 
     assert_eq!(1, query_without_static.iter(&world).count());
     assert_eq!(2, query_with_static.iter(&world).count());
@@ -341,27 +329,24 @@ fn mutate_remove_tag() {
 
 #[test]
 fn mutate_change_tag() {
-    let universe = Universe::new(None);
+    let universe = Universe::new();
     let mut world = universe.create_world();
 
-    let shared = (Model(5),).as_tags();
+    let shared = (Model(5),);
     let components = vec![
         (Pos(1., 2., 3.), Rot(0.1, 0.2, 0.3)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
         (Pos(4., 5., 6.), Rot(0.4, 0.5, 0.6)),
     ];
 
-    let entities = world.insert_from(shared, components).to_vec();
+    let entities = world.insert(shared, components).to_vec();
 
     let mut query_model_3 = <(Read<Pos>, Read<Rot>)>::query().filter(tag_value(&Model(3)));
     let mut query_model_5 = <(Read<Pos>, Read<Rot>)>::query().filter(tag_value(&Model(5)));
 
     assert_eq!(3, query_model_5.iter(&world).count());
     assert_eq!(0, query_model_3.iter(&world).count());
-
-    world.mutate_entity(*entities.get(1).unwrap(), |e| {
-        e.set_tag(Arc::new(Model(3)));
-    });
+    world.add_tag(*entities.get(1).unwrap(), Model(3));
 
     assert_eq!(2, query_model_5.iter(&world).count());
     assert_eq!(1, query_model_3.iter(&world).count());
